Auto-assign surfaces
This moves from explicit surface creation and assignment to generating
the superset of possible surfaces after reading the configuration and
applying the surfaces via surface options inside the styles.

It's PoC~ish and probably needs a bit of cleanup, since I only wanted
to touch as much as possible, and kept the string based lookup.